(chores) camel-core: code cleanup
Apply proper cast when needed

+++ core/camel-core/src/test/java/org/apache/camel/issues/SplitterParallelIssueTest.java
@@ -61,7 +61,7 @@ public class SplitterParallelIssueTest extends ContextTestSupport {
                             @Override
                             public void process(Exchange exchange) throws Exception {
                                 int num = exchange.getIn().getBody(int.class);
-                                final long sleep = num * delay;
+                                final long sleep = (long) num * delay;
                                 log.info("Sleep for {} ms", sleep);
                                 Thread.sleep(sleep);
                             }
